refactor(gamma): share the v2 flow mappings between the two modules
toStudioStep and toFlowExecution were byte-identical on both sides; gamma-ui-shared already carries this kind of code.
